nym-connect: function for kicking of the health check task
