+++ nym-connect/src-tauri/src/main.rs
@@ -49,7 +49,7 @@ fn main() {
             crate::operations::connection::status::get_connection_health_check_status,
             crate::operations::connection::status::get_connection_status,
             crate::operations::connection::status::get_gateway_connection_status,
-            crate::operations::connection::status::run_health_check,
+            crate::operations::connection::status::start_connection_health_check_task,
             crate::operations::directory::get_services,
             crate::operations::export::export_keys,
             crate::operations::window::hide_window,
