+++ nym-connect/src-tauri/src/operations/connection/mod.rs
@@ -1,3 +1,4 @@
 pub mod connect;
 pub mod disconnect;
+pub mod health_check;
 pub mod status;
